json grammar working | need to add actions

+++ week_01/examples/calculator/pic.cc
@@ -1,3 +1,3 @@
 #include "pic.hh"
 
-map<string,int> table;
+map<string,double> table;
